src/m365-agents-container/pyproject.toml
[project]
name = "azureai-foundry-streaming"
version = "0.1.0"
description = "Azure AI Foundry Streaming Agent Sample"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv",
    "aiohttp",
    "azure-identity",
    "agent-framework",
    "agent-framework-azure-ai",
    "azure-ai-agents",
    "microsoft-agents-activity",
    "microsoft-agents-hosting-core",
    "microsoft-agents-hosting-aiohttp",
    "microsoft-agents-authentication-msal",
]

[project.optional-dependencies]
dev = ["debugpy"]

[project.scripts]
azureai-foundry-streaming = "src.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
